Refactor ProductoService: pagination, search, DTOs
Introduce paginated search and DTO-driven service flow for Producto. Repository: add findBySearch(Pageable) and existsByNombre for name uniqueness checks. Model: change precio from Float to Double. Service: major refactor to use ProductoMapper, ProductoRequest/ProductoResponse, and PaginacionService; add obtenerPaginados and obtenerPaginadosConFiltro with sorting, implement guardar with uniqueness validation and audit, actualizarProducto with partial update & validation, buscarPorId returning a response, and borrarProducto keeping audit. Also add transactional handling and sorting helper; removed legacy DTO conversion utilities.

+++ src/main/java/org/example/topdeckapi/src/model/Producto.java
@@ -33,7 +33,7 @@ public class Producto {
     private String descripcion;
 
     @Column (name = "precio")
-    private Float precio;
+    private Double precio;
 
     @Column(name = "stock")
     private Integer stock;
